backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Supabase
supabase>=2.10.0
postgrest>=0.17.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Background Tasks
celery==5.3.6
redis==5.0.1

# Video Download (optional - only if you need crawler)
yt-dlp>=2024.10.7

# Subtitle Generation
assemblyai==0.26.0

# AI/ML (optional - only for quiz generation)
openai>=1.40.0
numpy>=2.0.0

# HTTP & WebSocket
httpx>=0.27.0,<0.28.0
websockets==12.0
tuspy>=1.0.3

# Utilities
python-dotenv==1.0.0
pytz==2024.1
aiofiles==23.2.1

# Monitoring & Logging
structlog==24.1.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.24.0
pytest-cov==4.1.0
